% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha_diversity_table}
\alias{alpha_diversity_table}
\title{Diversity indexes from endosphere samples of olive trees.}
\format{
A data frame with 24 rows and 6 variables: \describe{
\item{Observed}{observed richness} \item{Shannon}{Shannon diversity index}
\item{InvSimpson}{diversity index as inverse of simpson}
\item{Evenness}{Evenness index, calculated from Shannon}
\item{samples}{sample names}\item{location}{location of samples (either location1,
location2 or location3)} ... }
}
\source{
data from our laboratory.
}
\usage{
alpha_diversity_table
}
\description{
Data frame containing diversity indexes (Observed, Shannon, InvSimpson and
Evenness) from olive trees endosphere from three locations.
}
\keyword{datasets}
