% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancomLoop.R
\name{ancomloop}
\alias{ancomloop}
\title{ANCOM-BC analysis for multiple groups}
\usage{
ancomloop(
  input_object_phyloseq,
  grouping,
  ancom.options,
  out.unclassified = FALSE,
  tax.level
)
}
\arguments{
\item{input_object_phyloseq}{\link[phyloseq]{phyloseq-class}. For more
details, check \link[phyloseq]{distance} function.}

\item{grouping}{Character string. This is used as metadata variable for
reordering factors (which allows the function to loop over groups). It is
also passed to \code{group} and \link{formula} arguments in
\link[ANCOMBC]{ancombc} function.}

\item{ancom.options}{Further arguments to be passed to
\link[ANCOMBC]{ancombc}. They should be included as a list (see
\code{examples})}

\item{out.unclassified}{Whether to filter out unclassified taxa (taxa level
indicated with argument \code{tax.level}). Default is \code{FALSE}}

\item{tax.level}{Character string. Indicates at which level taxa should be
filtered out if are unclassified.}
}
\value{
Returns a list with as many arguments as groups are contained in
metadata "grouping" variable. Each element is named after the group which
is compared with all others and correspond to a data frame that includes
\link[ANCOMBC]{ancombc} parameters and the bias-adjusted abundance.
In the process, it prints the groups that are being compared (first group against others).
}
\description{
This function is a wrapper of \link[ANCOMBC]{ancombc}. When
having multiple groups to compare, it allows the user to automatically loop
over them and computes all possible combinations. Along with
\link[ANCOMBC]{ancombc} results, it appends to each table the taxonomy
classification.
}
\examples{

ancom_location<- ancomloop(location_phyloseq,grouping="location",
ancom.options=list(p_adj_method ="holm"),
out.unclassified=TRUE, tax.level="Genus")





}
