% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ASV_table_classified_raw}
\alias{ASV_table_classified_raw}
\title{ASV table from olive trees endosphere samples without cleaning.}
\format{
A data frame with 4,396 rows and 35   variables: \describe{
\item{ASV_seqs}{sequences infered with dada2}
\item{Kingdom}{classification} \item{Phylum}{classification}
\item{Class}{classification}
\item{Order}{classification} \item{Family}{classification}
\item{Genus}{classification} \item{ASV_names}{unique names for ASVs with
numbers} \item{L1-1}{From L1-1 to L1-8, samples of olive trees from location
\enumerate{
\item The same for L2 and L3} \item{MOCK-1}{from MOCK-1 to MOCK-3, the three
replicates from MOCK community sequencing}...}
}
}
\source{
data from our laboratory.
}
\usage{
ASV_table_classified_raw
}
\description{
Data frame containing ASV sequences, names, classification and abundance for
olive trees endosphere samples. This ASV table is raw, not cleaned from
eukaryotic sequences or MOCK community.
}
\keyword{datasets}
