% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betadispersion.R
\name{Betadispersion}
\alias{Betadispersion}
\title{Multivariate homogeneity of groups dispersions for several distance methods.}
\usage{
Betadispersion(
  data,
  formula,
  distances,
  permutest.options,
  betadisper.options,
  type = "Samples",
  ...
)
}
\arguments{
\item{data}{a  \link[phyloseq]{phyloseq-class}. For more details, check
\link[phyloseq]{distance} funtion.}

\item{formula}{Model formula to be passed to \code{group} argument from
\link[vegan]{betadisper}.}

\item{distances}{Character string including multiple distance methods to be
used. Further details to be found in \link[phyloseq]{distance}.}

\item{permutest.options}{Further arguments to be passed to
\link[vegan]{permutest}. They should be included as a list (see
\code{examples})}

\item{betadisper.options}{Further arguments to be passed to
\link[vegan]{betadisper}. They should be included as a list (see
\code{examples})}

\item{type}{character string with the type of comparison to used (sample-wise
or taxa-wise, with default c(anova.cca"Samples")). Check
\link[phyloseq]{distance} for further details.}

\item{...}{Further arguments to be passed to \link[phyloseq]{distance}
function from package \code{phyloseq}.}
}
\value{
Returns a list with to elements: \enumerate{ \item A data frame containing the
\code{tab} component of \link[vegan]{permutest} returns value, for every
distance used. This includes information of sources of variation, degrees of
freedom, sequential sums of squares, mean squares, \emph{F} statistics,
partial \emph{R-squared} and \emph{p-values}, based on \emph{N} permutations
\item A list with as many elements as distances were set. Each element
includes all information provided by an typical \link[vegan]{anova.cca} result
object.  }
}
\description{
This function is based on functions \link[phyloseq]{distance},
\link[vegan]{betadisper} and \link[vegan]{permutest} from \code{phyloseq}
and \code{vegan} packages. It calculates matrix distances on a
\link[phyloseq]{phyloseq-class} object (making use of
\link[phyloseq]{distance}) and then performs an analysis of multivariate
homogeneite of group dispersions (variances) based on the \code{formula}.
Then, it applies \link[vegan]{permutest} on these results. It allows the
user to wrap all these functions and loop over several distance methods
(Bray-Curtis, UniFrac, Weighted UniFrac...).
}
\examples{


betadisper_location<- Betadispersion(normalized_phyloseq,distances = c("bray",
"unifrac", "wunifrac"), formula = "location", betadisper.options= list( bias.adjust = TRUE))



}
