% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnT.R
\name{dunnT}
\alias{dunnT}
\title{Dunn's Test for multiple variables}
\usage{
dunnT(data, numberOfIndexes, formula, dunn.options, method = "BH")
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{formula}. Further details to be found in
\link[dunn.test]{dunn.test}.}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be taken as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column (corresponding to \code{g} argument in \code{dunn.test},
representing grouping vector or factor).}

\item{dunn.options}{Further arguments to be passed to
\link[dunn.test]{dunn.test}}

\item{method}{Method selected to adjust the \emph{p}-values for multiple
comparisons. See \link[dunn.test]{dunn.test} for more detailes about methods
and abbreviations. Default is set to  Benjamini-Hochberg adjustment ("BH")}
}
\value{
Returns a data frame with \code{dunn.test} results for all pairwise
comparisons, performed on each variable  (determined by
\code{numberOfIndexes}).For further details on parameters, check \link[dunn.test]{dunn.test}.
}
\description{
This function make use of \link[dunn.test]{dunn.test} function to
compute Dunn's test of multiple comparisons along groups (using rank sums)
and looping it along multiple variables of interest (e.g., alpha diversity
indexes).
}
\examples{
dunn_location <- dunnT(alpha_diversity_table,4,"location")


}
