% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskall.R
\name{kruskal.wallis}
\alias{kruskal.wallis}
\title{Kruskal-Wallis Rank Sum Test for multiple variables}
\usage{
kruskal.wallis(data, numberOfIndexes, formula, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{Formula}}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column (corresponding to \code{g} argument in
\link[stats]{kruskal.test}, representing grouping vector or factor).}

\item{...}{Further arguments passed to \link[stats]{kruskal.test}.}
}
\value{
Returns a data frame with p-values of the test for all variables
(determined by \code{numberOfIndexes}).
}
\description{
Computes a Kruskal-Wallis rank sum test for multiple variables. Input must be
a table with indexes as columns and samples as rows. Further columns should
include metadata.
}
\examples{

kruskal_location<-kruskal.wallis(alpha_diversity_table, 4, "location")

}
