% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Levene.R
\name{levene.test.alpha}
\alias{levene.test.alpha}
\title{Levene's test for multiple variables}
\usage{
levene.test.alpha(data, numberOfIndexes, formula, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to variables (e.g. alpha diversity indexes) and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{Formula}.}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column and specify the model.}

\item{...}{Further arguments to be passed to \link[car]{leveneTest}.}
}
\value{
returns a data frame with results from \link[car]{leveneTest} from package \code{car}
for each variable (determined by \code{numberOfIndexes}).Further details can
be found in \link[car]{leveneTest}
}
\description{
Test homogeneity of variance across groups, looping across
multiple variables (e.g. alpha diversity indexes).
}
\examples{

levene_location <- Levene.test.alpha(alpha_diversity_table, 4, "location")

}
