% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_phyloseq}
\alias{location_phyloseq}
\title{Phyloseq object containing classification, abundance, phylogenetic tree and
ASV sequences from olive trees endosphere samples.}
\format{
A phyloseq object with five elements: otu_table, tax_table, sam_data,
phy_tree and refseq.
}
\source{
data from our laboratory. This object has been produced form
ASV_final_table and tree data from this package.
}
\usage{
location_phyloseq
}
\description{
Phyloseq object.
}
\keyword{datasets}
