% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata table.}
\format{
Data frame with samples and location information.
}
\source{
produced with an internal R script.
}
\usage{
metadata
}
\description{
Metadata table containing information in order to group samples according to location.
}
\keyword{datasets}
