% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_composition}
\alias{mock_composition}
\title{MOCK community composition table.}
\format{
Data frame with genus classification.
}
\source{
ZymoBIOMICS Microbial Community Standard II (Log Distribution), ZYMO
RESEARCH, CA, USA.
}
\usage{
mock_composition
}
\description{
Table containing genus classification of MOCK community members.
}
\keyword{datasets}
