% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseAdonisFun.R
\name{PairwiseAdonisFun}
\alias{PairwiseAdonisFun}
\title{Pairwise multilevel comparison using adonis over multiple distance metrics}
\usage{
PairwiseAdonisFun(
  data,
  formula,
  distances,
  pval,
  p.adjust.m = "BH",
  pw.options,
  type = "samples",
  ...
)
}
\arguments{
\item{data}{\link[phyloseq]{phyloseq-class}. For more details, check
\link[phyloseq]{distance} funtion.}

\item{formula}{Column name of metadata to be passed to \code{factors} argument
from \link[pairwiseAdonis]{pairwise.adonis}.}

\item{distances}{Character string including multiple distance methods to be
used. Further details to be found in \link[phyloseq]{distance}.}

\item{pval}{\emph{p}-value threshold to filter results.}

\item{p.adjust.m}{Method selected to adjust the \emph{p}-values for multiple
comparisons. See \link[pairwiseAdonis]{pairwise.adonis} for more detailes
about methods and abbreviations. Default is set to  Benjamini-Hochberg
adjustment ("BH")}

\item{pw.options}{Further arguments to be passed to
\link[pairwiseAdonis]{pairwise.adonis}. They
should be included as a list (see \code{examples}).}

\item{type}{character string with the type of comparison to used (sample-wise
or taxa-wise, with default c(anova.cca"Samples")). Check
\link[phyloseq]{distance} for further details.}

\item{...}{Further arguments to be passed to \link[phyloseq]{distance}
function from package \code{phyloseq}.}
}
\value{
Returns a data frame with \link[pairwiseAdonis]{pairwise.adonis}
results for all pairwise comparisons, performed on each distance .For further
details on parameters, check \link[pairwiseAdonis]{pairwise.adonis}.
}
\description{
This is a wrapper function of
\link[pairwiseAdonis]{pairwise.adonis}. It takes a
\link[phyloseq]{phyloseq-class} object and calculates various distance
matrices. Then, with \link[pairwiseAdonis]{pairwise.adonis} it performs an
ANOVA analysis making use of \link[vegan]{adonis} and computes pairwise
comparisons if more than two groups are provided.
}
\examples{

#With no further arguments
pairwise_location<- PairwiseAdonisFun(normalized_phyloseq,distances = c("bray",
"unifrac", "wunifrac"),p.adjust.m="BH", formula = "location", pval=0.05)

#With further arguments
pairwise_location<- PairwiseAdonisFun(normalized_phyloseq,distances = c("bray",
"unifrac", "wunifrac"),p.adjust.m="BH", formula = "location", pval=0.05,
pw.options=list(perm=220))



}
