% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{Permanova}
\alias{Permanova}
\title{Permutational Multivariate Analysis of Variance along multiple distances
metrics}
\usage{
Permanova(data, formula, distances, type = "samples", adonis.options, ...)
}
\arguments{
\item{data}{a  \link[phyloseq]{phyloseq-class}. For more details, check
\link[phyloseq]{distance} funtion.}

\item{formula}{Model formula to be passed to \link[vegan]{adonis}.}

\item{distances}{Character string including multiple distance methods to be
used. Further details to be found in \link[phyloseq]{distance}.}

\item{type}{character string with the type of comparison to used (sample-wise
or taxa-wise, with default c("Samples")). Check \link[phyloseq]{distance}
for further details.}

\item{adonis.options}{Further arguments to be passed to \link[vegan]{adonis}. They
should be included as a list (see \code{examples})}

\item{...}{Further arguments to be passed to \link[phyloseq]{distance} function from
package \code{phyloseq}.}
}
\value{
Returns a list with to elements: \enumerate{ \item A data frame
containing the \code{aov.tab} component of resulting \link[vegan]{anova.cca}
from \link[vegan]{adonis} for every distance used. This includes information of
sources of variation, degrees of freedom, sequential sums of squares, mean
squares, \emph{F} statistics, partial \emph{R-squared} and \emph{p-values},
based on \emph{N} permutations \item A list with as many elements as
distances were set. Each element includes all information provided by an
typical \link[vegan]{anova.cca} result object.  }
}
\description{
This function is a modification of \link[vegan]{adonis} from
\code{vegan} package. It performs PERMANOVA analysis looping along a list of
distances metrics (Bray-Curtis, UniFrac, Weighted UniFrac...).
}
\examples{


permanova_location<- Permanova(normalized_phyloseq,distances = c("bray",
"unifrac", "wunifrac"), formula = "location", adonis.options= list(Permutations=200))


}
