% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro.R
\name{Shapiro}
\alias{Shapiro}
\title{Shapiro-Wilk Normality test for multiple variables.}
\usage{
Shapiro(data, numberOfIndexes, formula, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{Formula}.}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column and specify the model.}

\item{...}{Further arguments to be passed to \code{aov}.}
}
\value{
A data frame with p-values for each variable analyzed. \code{indexColumn} indicates variables names.
}
\description{
Computes a Shapiro-Wilk Normality test based on a model produced
with \code{aov}. It loops over multiple variables, determined by
\code{numberOfIndexes}.
}
\examples{
shapiro_rhizo<-Shapiro(alpha_indexes_rhizo, 4, "Management")



}
