% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Phylogenetic tree in \link[phyloseq]{phylo} format.}
\format{
Phyloseq tree object.
}
\source{
produced with an internal R script making used of ASV_final_table
data from this package.
}
\usage{
tree
}
\description{
Phylogenetic tree produced with MAFFT and FastTreeMP, further read into a
phyloseq tree object.
}
\keyword{datasets}
