% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tuckey.R
\name{Tukey.test}
\alias{Tukey.test}
\title{Multiple comparisons for multiple variables: balanced and unbalanced.}
\usage{
Tukey.test(data, numberOfIndexes, formula, balanced, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{Formula}.}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column and specify the model.}

\item{balanced}{whether to perform a balanced (TRUE) or unbalanced (FALSE)
test}

\item{...}{Further arguments to be passed to \code{HSD.test} or
\code{TukeyHSD}}
}
\value{
Returns a data frame with as many comparisons as groups, performed for
all variables (determined by \code{numberOfIndexes}). Details about
parameters should be check with \code{?TukeyHSD} and \code{?HSD.test}. If
\code{balanced=FALSE}, it returns the \code{comparison} element of
\code{HSD.test}.
}
\description{
Performs multiple comparisons in each variable. Variables should
be columns in \code{data}. This function is based on \code{TukeyHSD} and
\code{HSD.test} from packages \code{stats} and \code{agricolae} for multiple
comparisons. These are subsequently based on \code{aov} from \code{stats}
for model design.
}
\examples{

tukey_balanced<- Tukey.test(alpha_indexes_rhizo, numberOfIndexes = 4, formula = "Management", balanced=TRUE)
tukey_unbalanced<- Tukey.test(alpha_indexes_rhizo, numberOfIndexes = 4, formula = "Management", balanced=FALSE)

}
