% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnbalancedAnova.R
\name{UnbalancedAnova}
\alias{UnbalancedAnova}
\title{Performs unbalanced ANOVA on a \code{lm} model for several indexes}
\usage{
UnbalancedAnova(data, numberOfIndexes, formula, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{Formula}}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column and specify the model.}

\item{...}{Further arguments to be passed to \link[stats]{aov}.}
}
\value{
Returns a data frame with results from \code{Anova}. For more details, check
\link[car]{Anova} and \link[stats]{aov} from \code{car} and \code{stats} packages.
}
\description{
This function computes type-III analysis-of-variance on  a
\code{lm} model calculated with \link[stats]{aov} for several variables. Input must be
a table with indexes as columns and samples as rows. Further columns should
include metadata.
}
\examples{

unbalancedResult <- UnbalancedAnova(alpha_diversity_table, 4, "location")

}
