% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.R
\name{wilcoxon.test}
\alias{wilcoxon.test}
\title{Pairwise Wilcoxon Rank Sum Tests for multiple variables}
\usage{
wilcoxon.test(data, numberOfIndexes, formula, p.adjust.method, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used in argument \code{formula}. Further details to be found in
\link[stats]{pairwise.wilcox.test}.}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column (corresponding to \code{g} argument in
\code{pairwise.wilcox.test}, representing grouping vector or factor).}

\item{p.adjust.method}{method for adjusting p values (see
\link[stats]{p.adjust}). Can be abbreviated}

\item{...}{Further arguments passed to \code{pairwise.wilcox.test}.}
}
\value{
Returns a data frame with adjusted p-values for all pairwise
comparisons, performed on each variable  (determined by
\code{numberOfIndexes}).
}
\description{
Pairwise Wilcoxon Rank Sum Tests for multiple variables
}
\examples{
wilcoxon_rhizo_plots<- wilcoxon.test(alpha_indexes_rhizo, 4, "Plot", p.adjust.method="BH")

}
