% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnbalancedAnova.R
\name{UnbalancedAnova}
\alias{UnbalancedAnova}
\title{Performs unbalanced ANOVA on a \code{lm} model for several indexes}
\usage{
UnbalancedAnova(data, numberOfIndexes, formula, ...)
}
\arguments{
\item{data}{a data frame, columns corresponding to indexes and rows
corresponding to samples. Further columns should be included with metadata.
This is used for Formula as a grouping factor.}

\item{numberOfIndexes}{Integer corresponding to the number of indexes to
analyze. This will be take as column numbers by the function.}

\item{formula}{Metadata group name. This will group samples according to a
metadata column and specify the model.}

\item{...}{Further arguments to be passed to \code{aov}.}
}
\value{
Returns a data frame with results from \code{Anova}. For more details, check
\code{\link[stats]{aov}}  \code{\link[foo:bar]{foo::bar()}}
}
\description{
This function computes type-III analysis-of-variance on  a
\code{lm} model calculated with \code{aov} for several variables. Input must be
a table with indexes as columns and samples as rows. Further columns should
include metadata.
}
